medical_ssl/requirements.txt
torch
torchvision
timm
scikit-learn
pyyaml